Modified palos as a golang module
