--- alpine.json
@@ -1,15 +1,0 @@
-{
-    "name" : "alpine",
-    "enableIsolation" : true,
-    "enableResourceIsolation" : true,
-    "isolationOptions" : {
-        "enableUTS" : true,
-        "enablePID" : true,
-        "enableRoot" : true,
-        "enableNetNs" : false
-    },
-    "rootFs" : "/home/narasimha/alpine",
-    "nvidiaGpus" : [],
-    "ip" : "172.16.0.12",
-    "runtimeArgs" : ["/bin/ash"]
-}
