+++ src/github.com/Narasimha1997/pavlos/go.mod
@@ -1,0 +1,6 @@
+module github.com/Narasimha1997/pavlos
+
+go 1.14
+
+//requirements
+require github.com/jaypipes/ghw v0.6.1
